fix stling in items page
