+++ frontend/pos/package.json
@@ -12,6 +12,7 @@
     "bootstrap": "^4.5.2",
     "namor": "^2.0.2",
     "react": "^16.13.1",
+    "react-bootstrap": "^1.3.0",
     "react-dom": "^16.13.1",
     "react-modal": "^3.11.2",
     "react-redux": "^7.2.1",
