+++ frontend/pos/src/index.css
@@ -28,3 +28,38 @@ code {
 .btn {
   border: 0px !important;
 }
+
+.band-header {
+  background-color: #EBEBEB;
+  padding: 0.5rem 0rem;
+}
+
+th {
+  background-color: #2980b9;
+  color: #fff;
+  padding: 1rem;
+}
+
+.pagination {
+  width: 100%;
+  padding: 1rem;
+  display: flex;
+  justify-content: space-between;
+}
+
+.pagination__left > button {
+  background-color: var(--primary-clr);
+  color: white;
+  border: 0px;
+  border-radius: 0.2rem;
+}
+
+.pagination__right > span > input {
+  border: 1px solid #cbcaca;
+  border-radius: 0.2rem;
+}
+
+.pagination__right > span > select {
+  border: 1px solid #cbcaca;
+  border-radius: 0.2rem;
+}
