+++ frontend/pos/src/pages/ItemsPage/ItemsTable.js
@@ -1,0 +1,220 @@
+import React from "react";
+import styled from "styled-components";
+import { useTable, usePagination, useRowSelect } from "react-table";
+
+import makeData from "../../pages/ItemsPage/makeData";
+
+const Styles = styled.div`
+  padding: 1rem;
+`;
+
+const IndeterminateCheckbox = React.forwardRef(
+    ({ indeterminate, ...rest }, ref) => {
+        const defaultRef = React.useRef();
+        const resolvedRef = ref || defaultRef;
+
+        React.useEffect(() => {
+            resolvedRef.current.indeterminate = indeterminate;
+        }, [resolvedRef, indeterminate]);
+
+        return (
+            <>
+                <input type="checkbox" ref={resolvedRef} {...rest} />
+            </>
+        );
+    }
+);
+
+function ItemsTable({ columns, data }) {
+    // Use the state and functions returned from useTable to build your UI
+    const {
+        getTableProps,
+        getTableBodyProps,
+        headerGroups,
+        prepareRow,
+        page, // Instead of using 'rows', we'll use page,
+        // which has only the rows for the active page
+
+        // The rest of these things are super handy, too ;)
+        canPreviousPage,
+        canNextPage,
+        pageOptions,
+        pageCount,
+        gotoPage,
+        nextPage,
+        previousPage,
+        setPageSize,
+        selectedFlatRows,
+        state: { pageIndex, pageSize, selectedRowIds },
+    } = useTable(
+        {
+            columns,
+            data,
+        },
+        usePagination,
+        useRowSelect,
+        (hooks) => {
+            hooks.visibleColumns.push((columns) => [
+                // Let's make a column for selection
+                {
+                    id: "selection",
+                    // The header can use the table's getToggleAllRowsSelectedProps method
+                    // to render a checkbox
+                    Header: ({ getToggleAllPageRowsSelectedProps }) => (
+                        <div>
+                            <IndeterminateCheckbox {...getToggleAllPageRowsSelectedProps()} />
+                        </div>
+                    ),
+                    // The cell can use the individual row's getToggleRowSelectedProps method
+                    // to the render a checkbox
+                    Cell: ({ row }) => (
+                        <div>
+                            <IndeterminateCheckbox {...row.getToggleRowSelectedProps()} />
+                        </div>
+                    ),
+                },
+                ...columns,
+            ]);
+        }
+    );
+
+    // Render the UI for your table
+    return (
+        <>
+            <table className="table table-striped table-bordered table-sm table-hover" {...getTableProps()}>
+                <thead>
+                {headerGroups.map((headerGroup) => (
+                    <tr {...headerGroup.getHeaderGroupProps()}>
+                        {headerGroup.headers.map((column) => (
+                            <th {...column.getHeaderProps()}>{column.render("Header")}</th>
+                        ))}
+                    </tr>
+                ))}
+                </thead>
+                <tbody {...getTableBodyProps()}>
+                {page.map((row, i) => {
+                    prepareRow(row);
+                    return (
+                        <tr {...row.getRowProps()}>
+                            {row.cells.map((cell) => {
+                                return (
+                                    <td {...cell.getCellProps()}>{cell.render("Cell")}</td>
+                                );
+                            })}
+                        </tr>
+                    );
+                })}
+                </tbody>
+            </table>
+            {/*
+        Pagination can be built however you'd like.
+        This is just a very basic UI implementation:
+      */}
+            <div className="pagination">
+                <div className="pagination__left">
+                    <button onClick={() => gotoPage(0)} disabled={!canPreviousPage}>
+                        {"<<"}
+                    </button>{" "}
+                    <button onClick={() => previousPage()} disabled={!canPreviousPage}>
+                        {"<"}
+                    </button>{" "}
+                    <button onClick={() => nextPage()} disabled={!canNextPage}>
+                        {">"}
+                    </button>{" "}
+                    <button
+                        onClick={() => gotoPage(pageCount - 1)}
+                        disabled={!canNextPage}
+                    >
+                        {">>"}
+                    </button>{" "}
+                    <span>
+            Page{" "}
+                        <strong>
+              {pageIndex + 1} of {pageOptions.length}
+            </strong>{" "}
+          </span>
+                </div>
+                <div className="pagination__right">
+          <span>
+            Go to page:{" "}
+              <input
+                  type="number"
+                  defaultValue={pageIndex + 1}
+                  onChange={(e) => {
+                      const page = e.target.value ? Number(e.target.value) - 1 : 0;
+                      gotoPage(page);
+                  }}
+                  style={{ width: "100px" }}
+              />
+          </span>{" "}
+                    <select
+                        value={pageSize}
+                        onChange={(e) => {
+                            setPageSize(Number(e.target.value));
+                        }}
+                    >
+                        {[10, 20, 30, 40, 50].map((pageSize) => (
+                            <option key={pageSize} value={pageSize}>
+                                Show {pageSize}
+                            </option>
+                        ))}
+                    </select>
+                </div>
+            </div>
+        </>
+    );
+}
+
+function App() {
+    const columns = React.useMemo(
+        () => [
+            {
+                Header: "#",
+                accessor: "Number",
+            },
+            {
+                Header: "Name",
+                accessor: "Name",
+            },
+            {
+                Header: "Bar Code",
+                accessor: "barcode",
+            },
+            {
+                Header: "Category",
+                accessor: "category",
+            },
+            {
+                Header: "Cost Price",
+                accessor: "costprice",
+            },
+            {
+                Header: "Retail Price",
+                accessor: "retailprice",
+            },
+            {
+                Header: "Date Created",
+                accessor: "datecreated",
+            },
+            {
+                Header: "Last Modified",
+                accessor: "lastmodified ",
+            },
+            {
+                Header: "Actions",
+                accessor: "actions",
+            },
+        ],
+        []
+    );
+
+    const data = React.useMemo(() => makeData(100000), []);
+
+    return (
+        <Styles>
+            <ItemsTable columns={columns} data={data} />
+        </Styles>
+    );
+}
+
+export default App;
