+++ frontend/pos/src/pages/sales/sales.css
@@ -1,7 +1,4 @@
-.sales-header {
-  background-color: #EBEBEB;
-  padding: 0.5rem 0rem;
-}
+
 
 .items-table-header {
   background-color: #2980B9;
